pyproject.toml
[project]
name = "extraction-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.6.4",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.39.0",
    "fastapi>=0.136.1",
    "google-genai>=1.74.0",
    "httpx>=0.28.1",
    "openai>=2.33.0",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "uvicorn>=0.46.0",
]
